All the ServicesIpml

+++ projects/Project-Management-System-jdbc/src/Services/impl/Project_AllocationServiceImpl.java
@@ -1,40 +1,40 @@
 package Services.impl;
 
+import DB.DbException;
 import Entitys.Enums.Project_Status;
 import Entitys.Project_Allocation;
 
 import java.util.List;
-import Entitys.Company;
-import Entitys.Dao.CompanyDao;
+
 import Entitys.Dao.DeveloperDao;
 import Entitys.Dao.Factory.DaoFactory;
 import Entitys.Dao.ProjectDao;
 import Entitys.Dao.Project_AllocationDao;
 import Entitys.Developer;
 import Entitys.Enums.Developer_Status;
-import Entitys.Enums.Project_Status;
-import Entitys.Project;
-import Entitys.Project_Allocation;
 
-import Services.Exception.BussinesException;
+import Entitys.Project;
+
+
+import Services.Exception.BusinessException;
 import Services.model.Project_AllocationService;
 
 import java.time.LocalDate;
-import java.util.ArrayList;
+
 import java.util.HashMap;
-import java.util.List;
+
 import java.util.Map;
-import java.util.stream.Collectors;
+
 public class Project_AllocationServiceImpl implements Project_AllocationService{
-        private Project_AllocationDao projectAllocationDao= DaoFactory.creatproject_Allocation();
-        private DeveloperDao developerDao = DaoFactory.creatDeveloperJBBC();
-        private ProjectDao projectDao=DaoFactory.creatProject();
-        private CompanyDao companyDao =DaoFactory.creatCompanyJDBC();
+        private final Project_AllocationDao projectAllocationDao= DaoFactory.creatproject_Allocation();
+        private final DeveloperDao developerDao = DaoFactory.creatDeveloperJBBC();
+        private final ProjectDao projectDao=DaoFactory.creatProject();
+
         @Override
         public List<Project_Allocation> findByProject(Project project) {
             List<Project_Allocation> list =   projectAllocationDao.findByProject(project);
             if(list==null|| list.isEmpty()){
-                throw new BussinesException("Its empty");
+                throw new BusinessException("Its empty");
             }
             return list ;
         }
@@ -43,32 +43,30 @@ public class Project_AllocationServiceImpl implements Project_AllocationService{
         public List<Project_Allocation> findByDeveloper(Developer developer) {
             List<Project_Allocation> list =projectAllocationDao.findByDeveloper(developer);
             if(list==null|| list.isEmpty()){
-                throw new BussinesException("Its empty");
+                throw new BusinessException("Its empty");
             }
             return list ;
         }
 
         @Override
-        public List<Project_Allocation> findByPeriod(LocalDate start, LocalDate end) {
+        public List<Project_Allocation> findByHour(LocalDate start, LocalDate end) {
             if(start.isAfter(end) || end.isBefore(start)){
-                throw  new BussinesException("This datas cannot be used");
+                throw  new BusinessException("This datas cannot be used");
             }
             List<Project_Allocation> list= projectAllocationDao.findByHour(start,end);
 
             if(list==null|| list.isEmpty()){
-                throw new BussinesException("Its empty");
+                throw new BusinessException("Its empty");
             }
             return list ;
         }
 
         @Override
         public Map<Project_Status, List<Project_Allocation>> groupAllocationsByProjectStatus(Project project) {
-            if(projectAllocationDao.findByProject(project)==null){
-                throw new BussinesException("This project do not exist");
-            }
+
             List<Project_Allocation> list = projectAllocationDao.findByProject(project);
             if(list==null || list.isEmpty()){
-                throw  new BussinesException("No project with this status");
+                throw  new BusinessException("No project with this status");
             }
             Map<Project_Status, List<Project_Allocation>> map =new HashMap<>();
             map.put(project.getStatus(),list);
@@ -78,15 +76,19 @@ public class Project_AllocationServiceImpl implements Project_AllocationService{
 
         @Override
         public Map<Developer_Status, List<Project_Allocation>> groupAllocationsByDeveloperStatus(Developer developer) {
-            Map<Developer_Status, List<Project_Allocation>> map = new HashMap<>();
+            if (developer == null) {
+                throw new BusinessException("Developer cannot be null");
+            }
+
             List<Project_Allocation>list = projectAllocationDao.findByDeveloper(developer);
-            map.put(developer.getStatus(),list);
+
             if(list==null|| list.isEmpty()){
-                throw new BussinesException("this project allocation have 0 developers");
+                throw new BusinessException("this project allocation have 0 developers");
             }
-            if(map==null || map.isEmpty()){
-                throw new BussinesException("this project allocation have 0 developers with this satatus");
-            }
+
+            Map<Developer_Status, List<Project_Allocation>> map = new HashMap<>();
+            map.put(developer.getStatus(),list);
+
             return map;
         }
 
@@ -94,13 +96,17 @@ public class Project_AllocationServiceImpl implements Project_AllocationService{
         @Override
         public void update(Project_Allocation projectAllocation) {
             if(findById(projectAllocation.getId())==null){
-                throw new BussinesException("This project allocation do not exist");
+                throw new BusinessException("This project allocation do not exist");
             }
             projectAllocationDao.update(projectAllocation);
         }
 
         @Override
         public Project_Allocation deleteByID(Integer id) {
+            if(id==null){
+                throw  new DbException("Id cant be null");
+
+            }
             projectAllocationDao.deleteByID(id);
             return  projectAllocationDao.findById(id);
         }
@@ -109,7 +115,7 @@ public class Project_AllocationServiceImpl implements Project_AllocationService{
         public List<Project_Allocation> findAll() {
             List<Project_Allocation> list= projectAllocationDao.findAll();
             if(list==null|| list.isEmpty()){
-                throw new BussinesException("Its empty");
+                throw new BusinessException("Its empty");
             }
             return list ;
         }
@@ -118,7 +124,7 @@ public class Project_AllocationServiceImpl implements Project_AllocationService{
         public Project_Allocation findById(Integer id) {
             Project_Allocation pa = projectAllocationDao.findById(id);
             if(pa==null){
-                throw  new BussinesException("This id do not exist");
+                throw  new BusinessException("This id do not exist");
             }
             return pa;
         }
@@ -127,14 +133,14 @@ public class Project_AllocationServiceImpl implements Project_AllocationService{
         public void save(Project_Allocation projectAllocation) {
             Project proj = projectDao.findById(projectAllocation.getId_project());
             if(proj==null){
-                throw  new BussinesException("This Project Allocation do not exist");
+                throw  new BusinessException("This Project Allocation do not exist");
             }
-            if(developerDao.findById(projectAllocation.getId_developer())==null){
-                throw  new BussinesException("This Project Allocation do not exist");
+            if(projectAllocation.getId() != null && developerDao.findById(projectAllocation.getId_developer())==null){
+                throw  new BusinessException("This Project Allocation do not exist");
             }
 
             if (proj.getStatus() == Project_Status.Suspended || proj.getStatus() == Project_Status.Completed) { 
-                throw new BussinesException("You cant allocate to a project complete or suspended");
+                throw new BusinessException("You cant allocate to a project complete or suspended");
             }
 
             List<Project_Allocation> existingAllocations = projectAllocationDao.findByDeveloper(developerDao.findById(projectAllocation.getId_developer()));
@@ -147,7 +153,7 @@ public class Project_AllocationServiceImpl implements Project_AllocationService{
             }
 
             if (alreadyAllocated) {
-                throw new BussinesException("This developer is already allocated to this project");
+                throw new BusinessException("This developer is already allocated to this project");
             }
             projectAllocationDao.save(projectAllocation);
